pyproject.toml
[tool.poetry]
name = "be"
version = "0.1.0"
description = ""
authors = ["MichaelPosteraro15 <michaelposteraro65@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
coreapi = "^2.3.3"
pyyaml = "^6.0.2"
drf-yasg = "^1.21.8"
dj-rest-auth = "^7.0.0"
django-allauth = "^65.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
